backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.24.0
aiokafka>=0.8.0
kafka-python-ng>=2.0.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT>=2.10.1
email-validator>=2.0.0
